services/generation/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML/Audio
torch==2.1.0
torchaudio==2.1.0
transformers==4.35.0
accelerate==0.24.0
scipy==1.11.4
numpy==1.24.3

# Redis
redis==5.0.1

# Authentication
pyjwt==2.8.0

# ML extras
sentence-transformers==2.2.2
scikit-learn==1.3.2

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Utilities
python-dotenv==1.0.0
psutil==5.9.6
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0